partial integration testing

+++ s3-sink-connector/src/test/java/io/aiven/kafka/connect/s3/testutils/BucketAccessor.java
@@ -28,6 +28,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import com.amazonaws.services.s3.model.S3Object;
 import io.aiven.kafka.connect.common.config.CompressionType;
 
 import com.amazonaws.AmazonClientException;
@@ -37,6 +38,8 @@ import com.amazonaws.services.s3.model.MultiObjectDeleteException;
 import com.amazonaws.services.s3.model.S3ObjectSummary;
 import com.amazonaws.util.IOUtils;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import io.aiven.kafka.connect.common.source.NativeInfo;
+import org.apache.commons.io.function.IOSupplier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -58,26 +61,28 @@ public class BucketAccessor {
     }
 
     public final void removeBucket() {
-        final String[] chunk = s3Client.listObjects(bucketName)
-                .getObjectSummaries()
-                .stream()
-                .map(S3ObjectSummary::getKey)
-                .toArray(String[]::new);
-        if (chunk.length != 0) {
-            final var deleteObjectsRequest = new DeleteObjectsRequest(bucketName).withKeys(chunk);
-            try {
-                s3Client.deleteObjects(deleteObjectsRequest);
-            } catch (final MultiObjectDeleteException e) {
-                for (final var err : e.getErrors()) {
-                    LOGGER.warn(String.format("Couldn't delete object: %s. Reason: [%s] %s", err.getKey(),
-                            err.getCode(), err.getMessage()));
+        if (s3Client.doesBucketExistV2(bucketName)) {
+            final String[] chunk = s3Client.listObjects(bucketName)
+                    .getObjectSummaries()
+                    .stream()
+                    .map(S3ObjectSummary::getKey)
+                    .toArray(String[]::new);
+            if (chunk.length != 0) {
+                final var deleteObjectsRequest = new DeleteObjectsRequest(bucketName).withKeys(chunk);
+                try {
+                    s3Client.deleteObjects(deleteObjectsRequest);
+                } catch (final MultiObjectDeleteException e) {
+                    for (final var err : e.getErrors()) {
+                        LOGGER.warn(String.format("Couldn't delete object: %s. Reason: [%s] %s", err.getKey(),
+                                err.getCode(), err.getMessage()));
+                    }
+                } catch (final AmazonClientException e) {
+                    LOGGER.error("Couldn't delete objects: " + Arrays.stream(chunk).reduce(" ", String::concat)
+                            + e.getMessage());
                 }
-            } catch (final AmazonClientException e) {
-                LOGGER.error("Couldn't delete objects: " + Arrays.stream(chunk).reduce(" ", String::concat)
-                        + e.getMessage());
             }
+            s3Client.deleteBucket(bucketName);
         }
-        s3Client.deleteBucket(bucketName);
     }
 
     public final Boolean doesObjectExist(final String objectName) {
@@ -98,6 +103,10 @@ public class BucketAccessor {
                 .map(l -> l.split(","))
                 .map(fields -> decodeRequiredFields(fields, fieldsToDecode))
                 .collect(Collectors.toList());
+    }
+
+    public final IOSupplier<InputStream> getStream(final String blobName) {
+        return () ->  s3Client.getObject(bucketName, blobName).getObjectContent();
     }
 
     public final byte[] readBytes(final String blobName, final CompressionType compression) throws IOException {
@@ -147,5 +156,26 @@ public class BucketAccessor {
         Objects.requireNonNull(value, "value cannot be null");
 
         return new String(Base64.getDecoder().decode(value), StandardCharsets.UTF_8);
+    }
+
+    public final List<NativeInfo<S3Object, String>> getNativeInfo() {
+        List<S3ObjectSummary> lst = s3Client.listObjects(bucketName).getObjectSummaries();
+        return s3Client.listObjects(bucketName).getObjectSummaries().stream()
+                .map(objectSummary -> new NativeInfo<S3Object, String>() {
+            @Override
+            public S3Object getNativeItem() {
+                return s3Client.getObject(bucketName, getNativeKey());
+            }
+
+            @Override
+            public String getNativeKey() {
+                return objectSummary.getKey();
+            }
+
+            @Override
+            public long getNativeItemSize() {
+                return objectSummary.getSize();
+            }
+        }).collect(Collectors.toList());
     }
 }
